Separation of Concerns Time
In general, I restructured the code and project to better separate concerns. For example, database related work should be done in a db folder, the routes for Flashcards should be separated from the routes of Verbs, etc. Functionally this should work the same way, but now I can scale and a bit easier as my project continues to grow.
- created a backend folder (instead of src) so I can implement a React Frontend when the time comes.
- Added Routes for Flashcards and Verbs and removed them from being defined in `main.py`
- Created a db folder and stored all db related work in there.
- Created an llm folder for future llm work.
- moved the templates folder to the backend folder instead of the top level of the directory.
- created a `config.py` to read from my `.env` file and to add any other configs I want my project to have in the future.